+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `debugstack([thread, ][level, ntop, nbase, ...])` library API.
 - Added `strsplittable(delim, str[, limit])` library API.
 ### Changed
+- Fixed an issue where `secureexecuterange()` would not pass through additional arguments to the supplied function.
 - Fixed an issue where `strsplit()` would ignore the value supplied to the `limit` parameter.
 
 ## [v2.1]